refactor: TransactionsPage에서 테이블을 별도 컴포넌트로 분리
